+++ src/library/specex_psf_io_gauss_hermite_psf_fits_1.cc
@@ -1,4 +1,11 @@
+#include <fstream>
+#include <boost/algorithm/string.hpp>
+#include <harp.hpp>
+#include <specex_fits.h>
+#include <specex_trace.h>
+#include <specex_gauss_hermite_psf.h>
 
+using namespace std ;
 
 void write_gauss_hermite_psf_fits_version_1(const specex::GaussHermitePSF& psf, fitsfile* fp, int first_hdu) {
