+++ src/library/specex_psf_io_gauss_hermite_psf_fits_3.h
@@ -1,0 +1,6 @@
+#ifndef SPECEX_PSF_IO_GH3__H
+
+void write_gauss_hermite_psf_fits_version_3(const specex::GaussHermitePSF& psf, fitsfile* fp);
+void read_gauss_hermite_psf_fits_version_3(specex::PSF_p& psf, fitsfile* fp, int hdu, int first_bundle, int last_bundle);
+
+#endif
