+++ src/library/specex_psf_io_gauss_hermite_two_psf_fits_1.h
@@ -1,0 +1,5 @@
+#ifndef SPECEX_PSF_IO_GH2_1__H
+
+void write_gauss_hermite_two_psf_fits_version_1(const specex::GaussHermitePSF& psf, fitsfile* fp);
+
+#endif
