+++ src/library/specex_psf_io_gauss_hermite_two_psf_fits_2.h
@@ -1,0 +1,5 @@
+#ifndef SPECEX_PSF_IO_GH2_2__H
+
+void write_gauss_hermite_two_psf_fits_version_2(const specex::GaussHermitePSF& psf, fitsfile* fp);
+
+#endif
